Setup slime contribs if set before loading.
In order to allow a customizable set of slime contribs to be setup
when loading the helper, check whether slime-contribs has been set,
and if so pass that list to slime-setup; otherwise fall back on the
default '(slime-fancy) list.

slime-setup will use slime-contribs if it is called with no arguments,
and the default value for slime-contribs is '(slime-fancy), so this
could be done by calling slime-setup with no arguments. However, I'm
not certain if that is the case with all relevant versions of slime,
and we are already relying on the explicit-argument form, so that's
the conservative choice here.